flask-app/requirements.txt
pika==0.10.0
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Werkzeug==0.14.1
psycopg2-binary==2.7.7
